backend/auth/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg[binary]
alembic
python-jose[cryptography]
pydantic